Remove extra space after inlay hint for variables (#872)
Previously the inlay hint for variables would include a space at the
end. This meant that if you had a space between the variable and =, two
spaces would be displayed between the hint and =.

This looks pretty ugly imo, and means that with normal formatting there
would be a different amount of spaces for the inlay hints and the
explicit hints in the code.