fix: python versions < 3.4 broken by previous addition
My addition of quicker checking for python unittest modules breaks
python verison less than 3.4. `importlib.machinery.PathFinder` was added
in python 3.4 and will fail to import in verions prior to this.

This changes the code to only use the new check if the python version is
greater than or equal to 3.4. Note that `sys.version_info` was added in
version 2.0; unittest was added in python 2.1; older versions of pytest
support as low as python 2.7, so this change should be safe for any
python version supporting unit testing.